requirements.txt
flask
gunicorn
openai
requests
pyjwt
boto3
numpy
python-dotenv
